+++ gazebo/rendering/ScenePrivate.hh
@@ -281,7 +281,7 @@ namespace gazebo
       public: Heightmap *terrain;
 
       /// \brief The heightmap level of detail
-      public: double heightmapLOD = 0;
+      public: unsigned int heightmapLOD = 0u;
 
       /// \brief All the projectors.
       public: std::map<std::string, Projector *> projectors;
